feat: Implement shopping cart functionality with Cart and CartItem models; update UI to display cart items and total price

+++ app/src/main/java/com/nstut/fast_food_shop/util/Utils.java
@@ -4,7 +4,7 @@ import java.text.NumberFormat;
 import java.util.Locale;
 
 public class Utils {
-    public static String formatCurrency(int amount) {
+    public static String formatCurrency(double amount) {
         NumberFormat formatter = NumberFormat.getNumberInstance(new Locale("vi", "VN"));
         return formatter.format(amount) + "đ";
     }
